backend/pyproject.toml
[project]
name = "ad-reference-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "psycopg2-binary",
    "python-dotenv",
    "pydantic[email]",
    "pyjwt",
    "bcrypt",
    "httpx",
    "apscheduler",
    "playwright>=1.58.0",
]
